chore: common validate function and misc updates
This updates the validate schema function to be consistent with the
konnect-plugin-configuration-validation plugin. These two plugins may
get merged into one plugin so keeping them in sync with each other for a
potential helper function. The README and integration test were updated
also to add an example and fix the scoping of an integration run.